[Refactor] Refactor connector utils into common, Hive, and Iceberg helpers
