+++ be/src/connector/hive_chunk_sink.h
@@ -22,7 +22,6 @@
 #include "common/status.h"
 #include "common/thread/priority_thread_pool.hpp"
 #include "connector/partitioned_connector_chunk_sink.h"
-#include "connector/utils.h"
 #include "formats/column_evaluator.h"
 #include "formats/file_writer.h"
 #include "formats/io/async_flush_stream_poller.h"
