+++ be/src/connector/hive_utils.h
@@ -1,0 +1,40 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include <memory>
+#include <string>
+#include <vector>
+
+#include "common/statusor.h"
+#include "formats/column_evaluator.h"
+#include "types/type_descriptor.h"
+
+namespace starrocks::connector {
+
+class HiveUtils {
+public:
+    static StatusOr<std::string> make_partition_name(
+            const std::vector<std::string>& column_names,
+            const std::vector<std::unique_ptr<ColumnEvaluator>>& column_evaluators, Chunk* chunk,
+            bool support_null_partition);
+
+    static StatusOr<std::string> column_value(const TypeDescriptor& type_desc, const ColumnPtr& column, int idx);
+
+    template <typename T>
+    static StatusOr<std::string> format_decimal_value(T value, int scale);
+};
+
+} // namespace starrocks::connector
