+++ be/src/connector/iceberg_delete_sink.h
@@ -23,6 +23,7 @@
 #include "formats/file_writer.h"
 
 namespace starrocks {
+class PriorityThreadPool;
 class RuntimeState;
 
 namespace connector {
