+++ be/src/connector/partition_chunk_writer.h
@@ -23,7 +23,8 @@
 #include "compute_env/load_spill/load_chunk_spiller.h"
 #include "connector/common/connector_sink_commit.h"
 #include "connector/common/connector_sink_profile.h"
-#include "connector/utils.h"
+#include "connector/common/utils.h"
+#include "formats/column_evaluator.h"
 #include "formats/file_writer.h"
 #include "fs/fs_fwd.h"
