+++ be/test/connector_sink/file_chunk_sink_test.cpp
@@ -23,6 +23,7 @@
 
 #include "base/testutil/assert.h"
 #include "base/utility/defer_op.h"
+#include "connector/common/utils.h"
 #include "connector/partitioned_connector_chunk_sink.h"
 #include "connector_primitive/sink_memory_manager.h"
 #include "exec/exec_env.h"
